backend/requirements.txt
absl-py==1.0.0
asgiref==3.4.1
astunparse==1.6.3
bpemb==0.3.3
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
conllu==4.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
cycler==0.11.0
Cython==0.29.23
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
Django-Gtts==0.4
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
djoser==2.1.0
filelock==3.4.0
flair==0.10
flatbuffers==2.0
fonttools==4.28.3
ftfy==6.0.3
gast==0.4.0
gdown==3.12.2
gensim==4.1.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
gTTS==2.2.3
h5py==3.6.0
huggingface-hub==0.2.1
idna==3.3
importlib-metadata==4.8.2
itypes==1.2.0
Janome==0.4.1
Jinja2==3.0.3
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
konoha==4.6.5
langdetect==1.0.9
libclang==12.0.0
lxml==4.6.4
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.2.2
matplotlib-inline==0.1.3
matplotlib-venn==0.11.6
more-itertools==8.8.0
mpld3==0.3
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
overrides==3.1.0
packaging==21.3
pandas==1.1.5
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
segtok==1.5.10
sentencepiece==0.1.95
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
smart-open==5.2.1
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlitedict==1.7.0
sqlparse==0.4.2
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.22.0
tensorflow-text==2.7.3
termcolor==1.1.0
threadpoolctl==3.0.0
tokenizers==0.10.3
tomli==1.2.2
torch==1.10.0
tqdm==4.62.3
traitlets==5.1.1
transformers==4.12.5
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
Wikipedia-API==0.5.4
wrapt==1.13.3
zipp==3.6.0
